Got rid of extraneous stuff in tools/dumpBlastDB.py
dumpBlastDB.py was originally intended to dump the protein sequences from each of the genomes
in the MongoDB database into a FASTA file, index it with makeblastdb, then run a self-vs-self
blastp search.  It now just dumps the data out (as the name implies).  The rest will be done
by a bash script, which will then also call a python program to import the results in MongoDB.